reorganize all urls and views to follow app structure
