+++ ohmg/places/urls.py
@@ -1,9 +1,14 @@
-from django.urls import path
+from django.urls import path, register_converter
 
+from .converters import PlaceConverter
 from .views import (
     PlaceView,
+    Viewer,
 )
 
+register_converter(PlaceConverter, "place-slug")
+
 urlpatterns = [
-    path("<str:place_slug>/", PlaceView.as_view(), name="place"),
+    path("<place-slug:place>/", PlaceView.as_view(), name="place"),
+    path("viewer/<place-slug:place>/", Viewer.as_view(), name="viewer"),
 ]
